# backend/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AcompananteInput": {
        "additionalProperties": false,
        "description": "Fase 12E: acompañante nombrado (N:1 Reserva). Sin cuenta de usuario\npropia; responde a la pregunta abierta 6 de la Fase 12A (nombre+correo).\nCedula se omite por estar vacía en los datos reales.",
        "properties": {
          "correo": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Correo",
            "type": "string"
          },
          "nombre": {
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre",
            "type": "string"
          }
        },
        "required": [
          "nombre",
          "correo"
        ],
        "title": "AcompananteInput",
        "type": "object"
      },
...
      "EnsayoReservaResponse": {
        "description": "Ensayo asociado a una reserva (Fase 12E, N:1 Zona). Subset mínimo\npara la respuesta de reserva, mismo criterio que ZonaReservaResponse.",
        "properties": {
          "estado": {
            "$ref": "#/components/schemas/EstadoEntidad"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "zona_id": {
            "title": "Zona Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "nombre",
          "zona_id",
          "estado"
        ],
        "title": "EnsayoReservaResponse",
        "type": "object"
      },
...
      "EspacioReservaResponse": {
        "properties": {
          "capacidad": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidad"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoEntidad"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre",
          "estado"
        ],
        "title": "EspacioReservaResponse",
        "type": "object"
      },
...
      "EstadoEntidad": {
        "description": "Estado de entidades gestionables (espacios y recursos).",
        "enum": [
          "activo",
          "inactivo",
          "mantenimiento"
        ],
        "title": "EstadoEntidad",
        "type": "string"
      },
      "EstadoReserva": {
        "description": "Estados del ciclo de vida de una reserva.",
        "enum": [
          "esperando",
          "aprobada",
          "rechazada",
          "cancelada"
        ],
        "title": "EstadoReserva",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecursoReservaResponse": {
        "properties": {
          "capacidad": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidad"
          },
          "espacio": {
            "$ref": "#/components/schemas/EspacioReservaResponse"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoEntidad"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre",
          "estado",
          "espacio"
        ],
        "title": "RecursoReservaResponse",
        "type": "object"
      },
...
      "ReservaAcompananteResponse": {
        "properties": {
          "correo": {
            "title": "Correo",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre",
          "correo"
        ],
        "title": "ReservaAcompananteResponse",
        "type": "object"
      },
      "ReservaAsistioUpdate": {
        "description": "Fase 12D-bis: payload del endpoint dedicado `PUT /reservas/{id}/asistio`.\n\nSiempre se manda un valor concreto (`true`/`false`); no hay semántica de\n\"ausencia conserva\" como en `ReservaUpdate`.",
        "properties": {
          "asistio": {
            "title": "Asistio",
            "type": "boolean"
          }
        },
        "required": [
          "asistio"
        ],
        "title": "ReservaAsistioUpdate",
        "type": "object"
      },
      "ReservaCreate": {
        "additionalProperties": false,
        "properties": {
          "acompanantes": {
            "items": {
              "$ref": "#/components/schemas/AcompananteInput"
            },
            "title": "Acompanantes",
            "type": "array"
          },
          "asistentes": {
            "exclusiveMinimum": 0.0,
            "title": "Asistentes",
            "type": "integer"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "ensayo_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Ensayo Ids",
            "type": "array"
          },
          "fecha": {
            "format": "date",
            "title": "Fecha",
            "type": "string"
          },
          "hora_fin": {
            "format": "time",
            "title": "Hora Fin",
            "type": "string"
          },
          "hora_inicio": {
            "format": "time",
            "title": "Hora Inicio",
            "type": "string"
          },
          "recurso_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Recurso Ids",
            "type": "array"
          },
          "requiere_apoyo_auxiliar": {
            "default": false,
            "title": "Requiere Apoyo Auxiliar",
            "type": "boolean"
          },
          "tipo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TipoReserva"
              },
              {
                "type": "null"
              }
            ]
          },
          "tipo_solicitud": {
            "$ref": "#/components/schemas/TipoSolicitud",
            "default": "reserva_en_laboratorio"
          },
          "ubicacion_uso": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ubicacion Uso"
          },
          "zona_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Zona Ids",
            "type": "array"
          }
        },
        "required": [
          "fecha",
          "hora_inicio",
          "hora_fin",
          "asistentes"
        ],
        "title": "ReservaCreate",
        "type": "object"
      },
      "ReservaEstadoUpdate": {
        "properties": {
          "motivo": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "nuevo_estado": {
            "enum": [
              "aprobada",
              "rechazada",
              "cancelada"
            ],
            "title": "Nuevo Estado",
            "type": "string"
          }
        },
        "required": [
          "nuevo_estado"
        ],
        "title": "ReservaEstadoUpdate",
        "type": "object"
      },
      "ReservaResponse": {
        "properties": {
          "acompanantes": {
            "items": {
              "$ref": "#/components/schemas/ReservaAcompananteResponse"
            },
            "title": "Acompanantes",
            "type": "array"
          },
          "asistentes": {
            "title": "Asistentes",
            "type": "integer"
          },
          "asistio": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asistio"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "ensayo_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Ensayo Ids",
            "type": "array"
          },
          "ensayos": {
            "items": {
              "$ref": "#/components/schemas/EnsayoReservaResponse"
            },
            "title": "Ensayos",
            "type": "array"
          },
          "espacio": {
            "$ref": "#/components/schemas/EspacioReservaResponse"
          },
          "espacio_id": {
            "title": "Espacio Id",
            "type": "integer"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoReserva"
          },
          "fecha": {
            "format": "date",
            "title": "Fecha",
            "type": "string"
          },
          "hora_fin": {
            "format": "time",
            "title": "Hora Fin",
            "type": "string"
          },
          "hora_inicio": {
            "format": "time",
            "title": "Hora Inicio",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "motivo_rechazo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Rechazo"
          },
          "recurso_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Recurso Ids",
            "type": "array"
          },
          "recursos": {
            "items": {
              "$ref": "#/components/schemas/RecursoReservaResponse"
            },
            "title": "Recursos",
            "type": "array"
          },
          "requiere_apoyo_auxiliar": {
            "title": "Requiere Apoyo Auxiliar",
            "type": "boolean"
          },
          "tipo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TipoReserva"
              },
              {
                "type": "null"
              }
            ]
          },
          "tipo_solicitud": {
            "$ref": "#/components/schemas/TipoSolicitud"
          },
          "ubicacion_uso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ubicacion Uso"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioReservaResponse"
          },
          "usuario_id": {
            "title": "Usuario Id",
            "type": "integer"
          },
          "zona_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Zona Ids",
            "type": "array"
          },
          "zonas": {
            "items": {
              "$ref": "#/components/schemas/ZonaReservaResponse"
            },
            "title": "Zonas",
            "type": "array"
          }
        },
        "required": [
          "id",
          "usuario_id",
          "espacio_id",
          "fecha",
          "hora_inicio",
          "hora_fin",
          "estado",
          "asistentes",
          "tipo_solicitud",
          "requiere_apoyo_auxiliar",
          "created_at",
          "updated_at",
          "usuario",
          "espacio"
        ],
        "title": "ReservaResponse",
        "type": "object"
      },
      "ReservaUpdate": {
        "additionalProperties": false,
        "properties": {
          "acompanantes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AcompananteInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acompanantes"
          },
          "asistentes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asistentes"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "ensayo_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ensayo Ids"
          },
          "fecha": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha"
          },
          "hora_fin": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Fin"
          },
          "hora_inicio": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Inicio"
          },
          "recurso_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurso Ids"
          },
          "requiere_apoyo_auxiliar": {
            "default": false,
            "title": "Requiere Apoyo Auxiliar",
            "type": "boolean"
          },
          "tipo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TipoReserva"
              },
              {
                "type": "null"
              }
            ]
          },
          "tipo_solicitud": {
            "$ref": "#/components/schemas/TipoSolicitud",
            "default": "reserva_en_laboratorio"
          },
          "ubicacion_uso": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ubicacion Uso"
          },
          "zona_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zona Ids"
          }
        },
        "title": "ReservaUpdate",
        "type": "object"
      },
...
      "Rol": {
        "description": "Rol funcional de un usuario. Espejo del patrón\n`^(admin|gestor|usuario)$` de `app/schemas/usuario.py`.",
        "enum": [
          "usuario",
          "gestor",
          "admin"
        ],
        "title": "Rol",
        "type": "string"
      },
...
      "TipoReserva": {
        "description": "Tipo de reserva académica de `Reserva` (RN-012/RN-015, Fase 12D).\n\nNombre del campo en `Reserva`: `tipo` — el mismo nombre que ya citaba\nel código y la documentación como punto de extensión pendiente antes de\nexistir (services/reservas.py::validar_acceso_ps, models/README.md).\nTres valores exactos del roadmap aprobado, sin catálogo \"otro\":\nTRABAJO_INVESTIGACION (investigación), TRABAJO_GRADO (grado,\nproyectos de grado) y SERVICIO_DE_ENSAYO (ensayo, prestación de\nservicios). `servicio_de_ensayo` es el valor que el gate PS de la\nFase 12D exige para reservar un recurso marcado como PS.",
        "enum": [
          "trabajo_investigacion",
          "trabajo_grado",
          "servicio_de_ensayo"
        ],
        "title": "TipoReserva",
        "type": "string"
      },
      "TipoSolicitud": {
        "description": "Motivo de la solicitud (Fase B), del formulario real de solicitud de\nlaboratorios del ITM. Nombre del campo en `Reserva`: `tipo_solicitud`\n-- concepto DISTINTO de `tipo` (académico: trabajo_investigacion/\ntrabajo_grado/servicio_de_ensayo, `TipoReserva` arriba) y de\n`modalidad_reserva` de `Espacio` (`ModalidadEspacio`): los tres tienen\nnombres parecidos pero responden preguntas distintas.\n\nEl formulario real tiene 4 motivos; solo los 2 primeros viven acá.\nRESERVA_EN_LABORATORIO (dentro del laboratorio, el modelo de siempre) y\nRESERVA_FUERA_LABORATORIO (equipo usado fuera del laboratorio pero\ndentro de la sede) comparten la misma forma -- franja horaria de un\ndía -- así que ambos son ejes de `Reserva`. ORDEN_SALIDA (equipo fuera\nde la sede, rango de días) se agrega acá desde ya para no reabrir este\n`CheckConstraint` en la Fase C, pero solo existirá como fila real en\n`reservas` cuando `services/solicitudes.py` (Fase C) la materialice\ndía por día para lograr el bloqueo automático del calendario -- nunca\nllega directamente desde `ReservaCreate`. El cuarto motivo, MANO_OBRA,\nno usa ningún recurso ni bloquea nada: vive exclusivamente en la tabla\n`solicitudes_especiales` de la Fase C, nunca en `Reserva`.",
        "enum": [
          "reserva_en_laboratorio",
          "reserva_fuera_laboratorio",
          "orden_salida"
        ],
        "title": "TipoSolicitud",
        "type": "string"
      },
...
      "UsuarioReservaResponse": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "rol": {
            "$ref": "#/components/schemas/Rol"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "rol"
        ],
        "title": "UsuarioReservaResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ZonaReservaResponse": {
        "description": "Zona asociada a una reserva (Fase 12C-6). Misma base mínima que\n`ZonaResponse` sin timestamps/auditoría — suficiente para la respuesta\nde la reserva y los mensajes de notificación.",
        "properties": {
          "capacidad": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidad"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "espacio_id": {
            "title": "Espacio Id",
            "type": "integer"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoEntidad"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre",
          "espacio_id",
          "descripcion",
          "capacidad",
          "estado"
        ],
        "title": "ZonaReservaResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API para gestionar recursos y sus reservas",
    "title": "Gestión de Reservas de Recursos Institucionales",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reservas": {
      "get": {
        "operationId": "listar_reservas_endpoint_reservas_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReservaResponse"
                  },
                  "title": "Response Listar Reservas Endpoint Reservas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Listar Reservas Endpoint",
        "tags": [
          "reservas"
        ]
      },
      "post": {
        "operationId": "crear_reserva_endpoint_reservas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Crear Reserva Endpoint",
        "tags": [
          "reservas"
        ]
      }
    },
    "/reservas/mis-reservas": {
      "get": {
        "operationId": "listar_mis_reservas_endpoint_reservas_mis_reservas_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReservaResponse"
                  },
                  "title": "Response Listar Mis Reservas Endpoint Reservas Mis Reservas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Listar Mis Reservas Endpoint",
        "tags": [
          "reservas"
        ]
      }
    },
    "/reservas/{reserva_id}": {
      "delete": {
        "operationId": "eliminar_reserva_endpoint_reservas__reserva_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reserva_id",
            "required": true,
            "schema": {
              "title": "Reserva Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Eliminar Reserva Endpoint",
        "tags": [
          "reservas"
        ]
      },
      "patch": {
        "operationId": "actualizar_reserva_endpoint_reservas__reserva_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "reserva_id",
            "required": true,
            "schema": {
              "title": "Reserva Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Actualizar Reserva Endpoint",
        "tags": [
          "reservas"
        ]
      }
    },
    "/reservas/{reserva_id}/asistio": {
      "put": {
        "operationId": "marcar_asistencia_endpoint_reservas__reserva_id__asistio_put",
        "parameters": [
          {
            "in": "path",
            "name": "reserva_id",
            "required": true,
            "schema": {
              "title": "Reserva Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaAsistioUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Marcar Asistencia Endpoint",
        "tags": [
          "reservas"
        ]
      }
    },
    "/reservas/{reserva_id}/cancelar": {
      "put": {
        "operationId": "cancelar_reserva_usuario_endpoint_reservas__reserva_id__cancelar_put",
        "parameters": [
          {
            "in": "path",
            "name": "reserva_id",
            "required": true,
            "schema": {
              "title": "Reserva Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Cancelar Reserva Usuario Endpoint",
        "tags": [
          "reservas"
        ]
      }
    },
    "/reservas/{reserva_id}/estado": {
      "put": {
        "operationId": "cambiar_estado_endpoint_reservas__reserva_id__estado_put",
        "parameters": [
          {
            "in": "path",
            "name": "reserva_id",
            "required": true,
            "schema": {
              "title": "Reserva Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservaEstadoUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Cambiar Estado Endpoint",
        "tags": [
          "reservas"
        ]
      }
    },
...
  }
}